+++ src-docs/src/views/key_pad_menu/key_pad_menu_example.js
@@ -20,6 +20,10 @@ import KeyPadMenuItemButton from './key_pad_menu_item_button';
 const keyPadMenuItemButtonSource = require('!!raw-loader!./key_pad_menu_item_button');
 const keyPadMenuItemButtonHtml = renderToHtml(KeyPadMenuItemButton);
 
+import KeyPadBeta from './key_pad_beta';
+const keyPadBetaSource = require('!!raw-loader!./key_pad_beta');
+const keyPadBetaHtml = renderToHtml(KeyPadBeta);
+
 export const KeyPadMenuExample = {
   title: 'Key Pad Menu',
   sections: [{
@@ -54,5 +58,29 @@ export const KeyPadMenuExample = {
       </p>
     ),
     demo: <KeyPadMenuItemButton />,
+  }, {
+    title: 'Beta item',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: keyPadBetaSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: keyPadBetaHtml,
+    }],
+    text: (
+      <div>
+        <p>
+          If the item links to a module that is not GA (beta, lab, etc),
+          you can add a <EuiCode>betaBadgeLabel</EuiCode> and <EuiCode>betaBadgeTooltipContent</EuiCode> to
+          the card and it will properly create and position an <EuiCode>EuiBetaBadge</EuiCode>.
+        </p>
+        <p>
+          Supplying just a label will only show the first letter in the badge and supply the full label
+          to the tooltip. You can also pass an <EuiCode>iconType</EuiCode> to replace the letter only badge
+          and the label will still become the title.
+        </p>
+      </div>
+    ),
+    demo: <KeyPadBeta />,
   }],
 };