+++ src/components/key_pad_menu/key_pad_menu_item.js
@@ -2,8 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-const renderContent = (children, label) => (
+import { EuiBetaBadge } from '../../components/badge/beta_badge';
+
+import {
+  ICON_TYPES,
+} from '../icon';
+
+const renderContent = (children, label, betaBadgeLabel, betaBadgeTooltipContent, betaBadgeIconType) => (
   <div className="euiKeyPadMenuItem__inner">
+    {betaBadgeLabel &&
+      <span className="euiKeyPadMenuItem__betaBadgeWrapper">
+        <EuiBetaBadge
+          className="euiKeyPadMenuItem__betaBadge"
+          label={betaBadgeLabel}
+          iconType={betaBadgeIconType}
+          tooltipContent={betaBadgeTooltipContent}
+        />
+      </span>
+    }
+
     <div className="euiKeyPadMenuItem__icon">
       {children}
     </div>
@@ -17,10 +34,31 @@ const renderContent = (children, label) => (
 const commonPropTypes = {
   children: PropTypes.node.isRequired,
   label: PropTypes.node.isRequired,
+
+  /**
+   * Add a badge to the card to label it as "Beta" or other non-GA state
+   */
+  betaBadgeLabel: PropTypes.string,
+
+  /**
+   * Supply an icon type if the badge should just be an icon
+   */
+  betaBadgeIconType: PropTypes.oneOf(ICON_TYPES),
+
+  /**
+   * Add a description to the beta badge (will appear in a tooltip)
+   */
+  betaBadgeTooltipContent: PropTypes.node,
 };
 
-export const EuiKeyPadMenuItem = ({ href, label, children, className, ...rest }) => {
-  const classes = classNames('euiKeyPadMenuItem', className);
+export const EuiKeyPadMenuItem = ({ href, label, children, className, betaBadgeLabel, betaBadgeTooltipContent, betaBadgeIconType, ...rest }) => {
+  const classes = classNames(
+    'euiKeyPadMenuItem',
+    {
+      'euiKeyPadMenuItem--hasBetaBadge': betaBadgeLabel,
+    },
+    className
+  );
 
   return (
     <a
@@ -28,7 +66,7 @@ export const EuiKeyPadMenuItem = ({ href, label, children, className, ...rest })
       className={classes}
       {...rest}
     >
-      {renderContent(children, label)}
+      {renderContent(children, label, betaBadgeLabel, betaBadgeTooltipContent, betaBadgeIconType)}
     </a>
   );
 };
@@ -37,8 +75,14 @@ EuiKeyPadMenuItem.propTypes = ({ ...{
   href: PropTypes.string,
 }, ...commonPropTypes });
 
-export const EuiKeyPadMenuItemButton = ({ onClick, label, children, className, ...rest }) => {
-  const classes = classNames('euiKeyPadMenuItem', className);
+export const EuiKeyPadMenuItemButton = ({ onClick, label, children, className, betaBadgeLabel, betaBadgeTooltipContent, betaBadgeIconType, ...rest }) => {
+  const classes = classNames(
+    'euiKeyPadMenuItem',
+    {
+      'euiKeyPadMenuItem--hasBetaBadge': betaBadgeLabel,
+    },
+    className
+  );
 
   return (
     <button
@@ -47,7 +91,7 @@ export const EuiKeyPadMenuItemButton = ({ onClick, label, children, className, .
       className={classes}
       {...rest}
     >
-      {renderContent(children, label)}
+      {renderContent(children, label, betaBadgeLabel, betaBadgeTooltipContent, betaBadgeIconType)}
     </button>
   );
 };